buit ins in pipex

+++ srcs/17_pipex.c
@@ -6,7 +6,7 @@
 /*   By: anurtiag <anurtiag@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/20 07:13:42 by anurtiag          #+#    #+#             */
-/*   Updated: 2024/03/15 11:12:37 by anurtiag         ###   ########.fr       */
+/*   Updated: 2024/03/15 12:53:38 by anurtiag         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,24 +21,6 @@ void	ft_exit(int i)
 		exit(1);
 	}
 }
-
-// int	ft_built_in(char **argv, t_input **struct_input)
-// {
-// 	if (ft_strncmp(argv[0], "echo", ft_strlen(argv[0])) == 0 && ft_strncmp(argv[1], "-n", ft_strlen(argv[1])) == 0)
-// 		ft_echo(argv, 1);
-// 	else if(ft_strncmp(argv[0], "pwd", ft_strlen(argv[0])) == 0)
-// 		ft_pwd();
-// 	else if(ft_strncmp(argv[0], "cd", ft_strlen(argv[0])) == 0)
-// 		ft_cd(argv, struct_input);
-// 	else if(ft_strncmp(argv[0], "export", ft_strlen(argv[0])) == 0)
-// 		ft_export(argv[1], struct_input);
-// 	else if(ft_strncmp(argv[0], "unset", ft_strlen(argv[0])) == 0)
-// 		ft_unset(argv[1], struct_input);
-// 	else if(ft_strncmp(argv[0], "env", ft_strlen(argv[0])) == 0)
-// 		ft_print_var(*struct_input);
-// 	return (0);	
-// }
-
 
 void	freeall(char **str)
 {
@@ -56,14 +38,22 @@ void	freeall(char **str)
 	str = NULL;
 }
 
-int	pipex(t_var_parsed_table *cmd_list)
+int	pipex(t_input **struct_input)
 {
 	int		fd[2];
+	t_var_parsed_table	*cmd_list;
+	int		control;
 
-	
+	control = TRUE;
+	cmd_list = (*struct_input)->parsed_table;
+	if (!cmd_list->next)
+		ft_built_in(cmd_list->cmd_splited, struct_input, &control);
+	if (control == FALSE)
+		return(0);
 	fd[READ] = 0;
 	fd[WRITE] = 0;
-	ft_make_process(cmd_list, fd);
+	// ft_make_process(struct_input, fd);
+	ft_make_process(cmd_list, fd, struct_input);
 	if(access(".tempfile.txt", F_OK) == 0)
 		unlink(".tempfile.txt");
 	return (0);
